requirements.txt
OpenAI
langchain
streamlit
python-dotenv